Remove unnecessary gem install bundler
As I am somewhat unfamiliar with the Ruby ecosystem I was unaware of the
redundant step in my gh-pages-deploy workflow. The current `run: gem
install bundler jekyll` install bundler which is included in Ruby >2.6
and Jekyll. It then throws away all the work done installing Jekyll as
we set the config path for bundle and ask it to install Jekyll again
according to the Gemfile. Removing this line should increase workflow
speeds significantly.