Remove timestamp from posts
In the idea of the digital garden posts don't really have a publishing
date, they are revisited again and again and updated accordingly. The
only relevant timestamp for each post is the `last_modified`
one. `last_modified` could be set by git commits but for now it is set
manually.